Deduplicate command deployment into shared utility
Both ready.js and deploy-commands.js contained identical filesystem
scanning and REST registration logic. Extract into
src/utils/commandDeployer.js as the single source of truth.

The shared function also adds defensive coding that was missing from
both callers: skip non-directory entries via withFileTypes, wrap each
require() in try/catch so a malformed file doesn't abort deployment,
and guard toJSON() with a typeof check before calling it.
