+++ src/deploy-commands.js
@@ -1,37 +1,11 @@
-const { REST, Routes } = require('discord.js');
-const fs = require('fs');
-const path = require('path');
 require('dotenv').config();
-
-const commands = [];
-const foldersPath = path.join(__dirname, 'commands');
-const commandFolders = fs.readdirSync(foldersPath);
-
-for (const folder of commandFolders) {
-    const commandsPath = path.join(foldersPath, folder);
-    const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
-
-    for (const file of commandFiles) {
-        const filePath = path.join(commandsPath, file);
-        const command = require(filePath);
-        if ('data' in command && 'execute' in command) {
-            commands.push(command.data.toJSON());
-        }
-    }
-}
-
-const rest = new REST().setToken(process.env.DISCORD_TOKEN);
+const { deployCommands } = require('./utils/commandDeployer');
 
 (async () => {
     try {
-        console.log(`Started refreshing ${commands.length} application (/) commands.`);
-
-        const data = await rest.put(
-            Routes.applicationCommands(process.env.CLIENT_ID),
-            { body: commands },
-        );
-
-        console.log(`Successfully reloaded ${data.length} application (/) commands.`);
+        console.log('Started refreshing application (/) commands.');
+        const count = await deployCommands(process.env.CLIENT_ID, process.env.DISCORD_TOKEN);
+        console.log(`Successfully reloaded ${count} application (/) commands.`);
     } catch (error) {
         console.error(error);
     }
